[EditItem] separate logic and ui. [Login] bugfix
